[feat/#1178] 에러 다이얼로그 컴포넌트화
